+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ### `testlib`
 
+* Add `Test.Cardano.Ledger.Core.Binary.Huddle` with `huddleCodecSpecFull` and `huddleCodecSpecNoAnn` harnesses
 * Add `TestBlockHeader` and `mkTestBlockHeaderNoNonce` as a replacement to deprecated `BHeaderView` and `makeHeaderView`.
 * Modify `ToExpr` instance for `Mismatch` to display type-level `r` parameter using `Typeable`